rafactor: 자주 사용하는 totalTime 계산 utils 분리
