+++ src/utils/totalTime.js
@@ -1,0 +1,9 @@
+export const formatTotalTime = (duration) => {
+  const match = duration.match(/PT(?:(\d+)H)?(?:(\d+)M)?(?:(\d+)S)?/);
+  const hours = parseInt(match[1] || "0", 10);
+  const minutes = parseInt(match[2] || "0", 10);
+  const seconds = parseInt(match[3] || "0", 10);
+  const totalSeconds = hours * 3600 + minutes * 60 + seconds;
+
+  return totalSeconds;
+};
